decomp: fix sub_0802C62C prototype (u8 -> s8), sub_0802BC14 to 50% (56/112)
sub_0802C62C is still unmatched (naked), so its guessed prototype was
safe to correct: retail sign-extends the byte argument via asrs, not
the unsigned lsrs a u8 param implies. Fixing it to s8 also improved
the caller sub_0802BC14 from 41.1% to 50.0% by removing a redundant
re-cast at the call site.

+++ src/wip/sub_0802BC14.c
@@ -10,7 +10,7 @@ s32 sub_0802BC14(u16 a)
     s8 sa;
 
     sa = (s8)a;
-    count = sub_0802C62C((u8)sa);
+    count = sub_0802C62C(sa);
     limit = _0802BA7C(sa);
     sub_08067B98((void *)0x0833BE30, (s16)a, count, limit);
     if (count >= limit)
